// React/app/GerAdv_TS/CargosEsc/Interfaces/interface.CargosEsc.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
// copyright © 2000-2025 Menphis - Sistemas Inteligentes
// This file is part of the Source Genesys project                     
'use client';

export interface ICargosEsc {
// 202501251
    id: number;
 
	percentual: number,
	nome: string,
	classificacao: number,
}

export interface ICargosEscFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface ICargosEscIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}